move stuff to schema helpers
